// src/enums/contextEnum.ts
enum contextEnum {
  DATA = 'Dados',
  GROUP = 'Grupos',
  TYPES = 'Tipos',
  USERS = 'Usuários',
  RULES = 'Regras de processamento',
  ALERTS = 'Alertas',
  COMPANY = 'Empresas',
  LEARNING = 'Aprendizagem de máquina',
  SERVICES = 'Ordens de Serviço',
  DASHBOARD = 'Dashboards',
  THEMES = 'Temas',
  MANAGE_DASHBOARDS = 'Gerenciar Gráficos',
}

export default contextEnum